refactor(BE): 엔티티 패키지 재배치 service/ → entity/ (#41)
JPA 엔티티가 service/ 패키지에 혼재해 레이어드 아키텍처를 위반하던 문제를 해결.
User, Review, Restaurant, RestaurantReviewSummary를 entity/ 패키지로 이동하고
이를 참조하는 모든 파일의 import 경로를 업데이트.
