+++ backend/src/main/java/com/edurican/enchelinbe/entity/Restaurant.java
@@ -1,4 +1,4 @@
-package com.edurican.enchelinbe.service;
+package com.edurican.enchelinbe.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
