+++ backend/src/main/java/com/edurican/enchelinbe/repository/RestaurantRepository.java
@@ -1,6 +1,6 @@
 package com.edurican.enchelinbe.repository;
 
-import com.edurican.enchelinbe.service.Restaurant;
+import com.edurican.enchelinbe.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
