+++ backend/src/main/java/com/edurican/enchelinbe/repository/RestaurantReviewSummaryRepository.java
@@ -1,6 +1,6 @@
 package com.edurican.enchelinbe.repository;
 
-import com.edurican.enchelinbe.service.RestaurantReviewSummary;
+import com.edurican.enchelinbe.entity.RestaurantReviewSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
