+++ backend/src/main/java/com/edurican/enchelinbe/service/ReviewService.java
@@ -10,6 +10,9 @@ import com.edurican.enchelinbe.dto.UserStatsResponse;
 import com.edurican.enchelinbe.enums.EntityStatus;
 import com.edurican.enchelinbe.repository.RestaurantRepository;
 import com.edurican.enchelinbe.repository.ReviewRepository;
+import com.edurican.enchelinbe.entity.Restaurant;
+import com.edurican.enchelinbe.entity.Review;
+import com.edurican.enchelinbe.entity.User;
 import com.edurican.enchelinbe.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
