+++ backend/src/main/java/com/edurican/enchelinbe/service/ReviewSummaryService.java
@@ -7,6 +7,9 @@ import com.edurican.enchelinbe.common.exception.ErrorCode;
 import com.edurican.enchelinbe.dto.ReviewSummaryResponse;
 import com.edurican.enchelinbe.repository.RestaurantRepository;
 import com.edurican.enchelinbe.repository.RestaurantReviewSummaryRepository;
+import com.edurican.enchelinbe.entity.Restaurant;
+import com.edurican.enchelinbe.entity.RestaurantReviewSummary;
+import com.edurican.enchelinbe.entity.Review;
 import com.edurican.enchelinbe.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
