+++ backend/src/main/java/com/edurican/enchelinbe/service/UserService.java
@@ -3,6 +3,7 @@ package com.edurican.enchelinbe.service;
 import com.edurican.enchelinbe.common.exception.BusinessException;
 import com.edurican.enchelinbe.common.exception.ErrorCode;
 import com.edurican.enchelinbe.dto.UserProfileResponse;
+import com.edurican.enchelinbe.entity.User;
 import com.edurican.enchelinbe.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
